requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.1
urllib3==1.26.15
standard-imghdr==3.13.0
redis==6.2.0
vk-api==11.9.9